requirements.txt
pyserial==3.5
pyusb==1.3.1
hidapi==0.14.0.post4
